Added /temp_downloads to gitignore
This change adds the /temp_downloads directory to the .gitignore file. By adding this line, any files and subdirectories within the temp_downloads folder will be ignored by Git and will not be tracked in the version control system. This is a common practice to prevent temporary or downloaded files, which are not part of the source code, from being accidentally committed to the repository. This helps in keeping the repository clean and focused on the essential project files.